Make RenderRecord public
